refactor(account): move token command from main level to under account
