+++ internal/commands/account/token/create_test.go
@@ -1,4 +1,4 @@
-package tokens
+package token
 
 import (
 	"testing"
