+++ internal/commands/account/token/delete_test.go
@@ -1,4 +1,4 @@
-package tokens
+package token
 
 import (
 	"testing"
