+++ internal/commands/account/token/list_test.go
@@ -1,4 +1,4 @@
-package tokens
+package token
 
 import (
 	"bytes"
