+++ internal/commands/account/token/token.go
@@ -1,11 +1,11 @@
-package tokens
+package token
 
 import (
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands"
 )
 
-// BaseTokensCommand creates the base 'token' command
-func BaseTokensCommand() commands.Command {
+// BaseTokenCommand creates the base 'token' command
+func BaseTokenCommand() commands.Command {
 	return &tokensCommand{commands.New("token", "Manage tokens")}
 }
